Wrap error for aborting selection
