+++ Makefile
@@ -64,7 +64,7 @@ install: ## Install rps to /usr/local/bin and config to /usr/local/share/rps
 	install -m644 ./config.yaml ${DEST_DIR}${DATA_DIR}/$(BIN_NAME)/
 
 .PHONY: install_no_config 
-install_no_config: all ## Install rps to /usr/local/bin
+install_no_config: ## Install rps to /usr/local/bin
 	install -d ${DEST_DIR}${BIN_DIR}
 	install -m755 $(DIST_DIR)/$(BIN_NAME) ${DEST_DIR}${BIN_DIR}/
