+++ internal/prompt/fzf.go
@@ -2,6 +2,7 @@ package prompt
 
 import (
 	"context"
+	"errors"
 	"fmt"
 
 	"github.com/alexmerren/rps/internal/github/repository"
@@ -9,7 +10,13 @@ import (
 )
 
 func NewFzfPrompt(ctx context.Context, repositories []*repository.Repository) (int, error) {
-	return fzf.Find(repositories, func(index int) string {
+	index, err := fzf.Find(repositories, func(index int) string {
 		return fmt.Sprintf("%s/%s", repositories[index].GetOwner(), repositories[index].GetName())
 	})
+
+	if errors.Is(err, fzf.ErrAbort) {
+		return 0, ErrPromptInterrupted
+	}
+
+	return index, err
 }
